✨ Feat: delete images from db & cloud

+++ server/src/services/admin/product.services.js
@@ -1,7 +1,11 @@
+const Image = require("../../models/images");
 const Product = require("../../models/product");
+const { deleteImage, deleteFolder } = require("../../utils/cloudinary");
 
 const getProductInfoDB = async ({ slug }) => {
-  return await Product.findOne({ slug }).populate("category");
+  const product = await Product.findOne({ slug }).populate("category");
+  const images = await Image.find({ product_id: product._id });
+  return { ...product._doc, images };
 };
 
 const getProductsDB = async () => {
@@ -18,7 +22,34 @@ const updateProductDB = async (id, data) => {
 };
 
 const deleteProductDB = async (id) => {
+  // product folder path
+  const folderPath = `ssgt/${id}`;
+
+  // delete from cloudinary
+  const result = await deleteFolder(folderPath);
+  if (!result.success) return null;
+
+  // delete from db
+  await Image.deleteMany({ product_id: id });
+
+  // delete product
   return await Product.findByIdAndDelete(id);
 };
 
-module.exports = { getProductInfoDB, getProductsDB, createProductDB, updateProductDB, deleteProductDB };
+const addProductImagesDB = async (data) => {
+  return await Image.insertMany(data);
+};
+
+const deleteProductImageDB = async (public_id) => {
+  return await Image.findOneAndDelete({ public_id });
+};
+
+module.exports = {
+  getProductInfoDB,
+  getProductsDB,
+  createProductDB,
+  updateProductDB,
+  deleteProductDB,
+  addProductImagesDB,
+  deleteProductImageDB,
+};
